Moves Schema and Results related logic to seperate individual files
